sys-kernel: enable throttled background buffered writeback
This feature prevents the background writeback thread from stalling I/O,
e.g. reads.

For instance, without this option enabled, it's possible to stall reads
setting a writeback dirty threshold (/proc/sys/vm/dirty_writeback_centisecs)
sufficiently long.

Quoting the author of the feature:

"""
[..] If you are sick of Linux bogging down when buffered
writes are happening, then this is for you, laptop or server. The
patchset is fully stable, I have not observed problems. It passes full
xfstest runs, and a variety of benchmarks as well. It works equally well
on blk-mq/scsi-mq, and "classic" setups.
"""

"""
This feature adds a simple blk-wb code that keeps limits how much buffered
writeback we keep in flight on the device end. [..] it should be pretty
much auto-tuning.
""

Signed-off-by: Ezequiel Garcia <ezequiel@vanguardiasur.com.ar>